requirements.txt
# Core Flask dependencies
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7

# PDF processing
PyPDF2==3.0.1

# Google Gemini AI
google-generativeai==0.3.2

# LangChain and embeddings
langchain==0.0.350
langchain-google-genai==0.0.6
langchain-community==0.0.12

# Vector database
chromadb==0.4.18

# Text processing and utilities
tiktoken==0.5.2
pydantic==2.5.2

# Optional: Better PDF processing (alternative to PyPDF2)
# Uncomment if needed for better text extraction
# pypdf==3.17.4
# pdfplumber==0.10.3

# Development dependencies (optional)
# python-dotenv==1.0.0

# Additional utilities
requests==2.31.0
numpy==1.24.4